+++ src/components/AccountList.js
@@ -1,0 +1,43 @@
+import React, { useContext } from "react";
+import { useTranslation } from "react-i18next";
+import { formateDecimal } from "../utils/Utils";
+import StoreContext from "../store/store-context";
+
+const AccountTile = (props) => {
+  return (
+    <ul className="d-flex justify-content-between market-order-item market-balance">
+      <li>{props.account?.currency || "--"}</li>
+      <li>{formateDecimal(props.account?.total, { decimalLength: 8 })}</li>
+      <li>{formateDecimal(props.account?.balance, { decimalLength: 8 })}</li>
+      <li>{formateDecimal(props.account?.locked, { decimalLength: 8 })}</li>
+    </ul>
+  );
+};
+
+const AccountList = (props) => {
+  const storeCtx = useContext(StoreContext);
+  const { t } = useTranslation();
+  return (
+    <div className="account-list">
+      <ul className="d-flex justify-content-between market-order-item market-order__title market-balance">
+        <li>{t("currency")}</li>
+        <li>{t("totalBal")}</li>
+        <li>{t("availBal")}</li>
+        <li>{t("frozenBal")}</li>
+      </ul>
+      <ul className="order-list scrollbar-custom">
+        {storeCtx.selectedTicker?.instId && storeCtx.accounts?.accounts ? (
+          storeCtx.selectedTicker.instId
+            .split("-")
+            ?.map((ccy) => (
+              <AccountTile account={storeCtx.accounts.accounts[ccy]} />
+            ))
+        ) : (
+          <div></div>
+        )}
+      </ul>
+    </div>
+  );
+};
+
+export default AccountList;
